RESTful Web Project Creation
Created drugviser-war, which is the web project for handling all the REST
requests, now is possible to access it just by deploying drugviser-ear
(the war is packed inside the ear). URL to access the server resources:
http://localhost:8080/ws/services/{service.name}. You can use "medicine"
Service to tryout the functionality.

Created drugviser-entity project, containing all the Hibernate/JPA Entity.
It may be or may be not used for JSON Pojo's in the future, depending on
how they'll be serialized by the WildFly JAX-RS ObjectMapper.